+++ CHANGELOG.md
@@ -78,6 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed QuestItem interaction with Lectern, Campfire and Composter
 - fixed journal update after magic inventory is closed
 - fixed lever event not toggling the lever
+- fixed ConcurrentModificationException in PlayerData
 ### Security
 - fixed issue, where objectives that count things are out of sync with the database. This has also affected BungeeCord support